+++ src/main/java/net/frozenblock/wilderwild/block/TumbleweedBlock.java
@@ -19,6 +19,8 @@
 package net.frozenblock.wilderwild.block;
 
 import com.mojang.serialization.MapCodec;
+import net.frozenblock.wilderwild.block.impl.SnowloggingUtils;
+import net.frozenblock.wilderwild.config.BlockConfig;
 import net.frozenblock.wilderwild.entity.Tumbleweed;
 import net.frozenblock.wilderwild.registry.RegisterEntities;
 import net.frozenblock.wilderwild.registry.RegisterSounds;
@@ -94,7 +96,7 @@ public class TumbleweedBlock extends BushBlock implements SimpleWaterloggedBlock
 
 	@Nullable
 	public BlockState getStateForPlacement(@NotNull BlockPlaceContext context) {
-		BlockState state = super.getStateForPlacement(context);
+		BlockState state =  SnowloggingUtils.getSnowPlacementState(super.getStateForPlacement(context), context);
 		if (state != null) {
 			FluidState fluidState = context.getLevel().getFluidState(context.getClickedPos());
 			boolean waterlogged = fluidState.getType() == Fluids.WATER;
@@ -139,5 +141,8 @@ public class TumbleweedBlock extends BushBlock implements SimpleWaterloggedBlock
 	@Override
 	protected void createBlockStateDefinition(@NotNull StateDefinition.Builder<Block, BlockState> builder) {
 		builder.add(WATERLOGGED);
+		if (BlockConfig.get().snowloggingConfig.snowlogging) {
+			builder.add(SnowloggingUtils.SNOW_LAYERS);
+		}
 	}
 }